+++ src/60_tools/ft_add_to_str.c
@@ -39,8 +39,8 @@ void	ft_add_to_str(t_data *data, char **str, char *to_add, int position)
 	new_str = ft_malloc(data, sizeof(char) * (new_size + 1));
 	if (!new_str)
 		ft_exit(data);
-	if (position > ft_strlen(*str))
-		position = ft_strlen(*str);
+	if (position > (int)ft_strlen(*str))
+		position = (int)ft_strlen(*str);
 	ft_ats_copy_str(position, str, new_str, to_add);
 	ft_free(data, *str);
 	*str = new_str;
