+++ src/60_tools/ft_rm_from_str.c
@@ -11,8 +11,8 @@ void	ft_rm_from_str(t_data *data, char **s, int rm_len, int position)
 	new_s = ft_malloc(data, sizeof(char) * (new_size + 1));
 	if (!new_s)
 		ft_exit(data);
-	if (position > ft_strlen(*s))
-		position = ft_strlen(*s);
+	if (position > (int)ft_strlen(*s))
+		position = (int)ft_strlen(*s);
 	i = 0;
 	while (i < position)
 	{
